+++ appveyor/appveyor.ps1
@@ -2,8 +2,8 @@ $env:JAVA_HOME="C:\Program Files\Java\jdk1.8.0"
 $env:PATH="$($env:JAVA_HOME)\bin;$($env:PATH)"
 $env:CCM_PATH="C:\Users\appveyor\ccm"
 $env:CASSANDRA_VERSION=$env:cassandra_version
-$env:EVENT_LOOP_MANAGER="async"
-$env:SIMULACRON_JAR="C:\Users\appveyor\simulacron-standalone-0.7.0.jar"
+$env:EVENT_LOOP_MANAGER="asyncore"
+$env:SIMULACRON_JAR="C:\Users\appveyor\simulacron-standalone-0.5.0.jar"
 
 python --version
 python -c "import platform; print(platform.architecture())"