+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Features
 * Include hash of result set metadata in prepared stmt id (PYTHON-808)
 * Add NO_COMPACT startup option (PYTHON-839)
 * Add new exception type for CDC (PYTHON-837)
+* Add asyncio reactor (Python 3.6+ only) (PYTHON-507)
 
 Bug Fixes
 ---------